Move status models
