--- botx/bot/models/status/bot_menu.py
@@ -1,3 +1,0 @@
-from typing import Dict, NewType
-
-BotMenu = NewType("BotMenu", Dict[str, str])
