--- botx/bot/models/status/recipient.py
@@ -1,30 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional
-from uuid import UUID
-
-from botx.models.enums import ChatTypes
-from botx.models.message.incoming_message import IncomingMessage
-
-
-@dataclass
-class StatusRecipient:
-    bot_id: UUID
-    huid: UUID
-    ad_login: Optional[str]
-    ad_domain: Optional[str]
-    is_admin: Optional[bool]
-    chat_type: ChatTypes
-
-    @classmethod
-    def from_incoming_message(
-        cls,
-        incoming_message: IncomingMessage,
-    ) -> "StatusRecipient":
-        return StatusRecipient(
-            bot_id=incoming_message.bot_id,
-            huid=incoming_message.sender.huid,
-            ad_login=incoming_message.sender.ad_login,
-            ad_domain=incoming_message.sender.ad_domain,
-            is_admin=incoming_message.sender.is_chat_admin,
-            chat_type=incoming_message.chat.type,
-        )
